+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/spend
 
+## 0.12.135
+
+### Patch Changes
+
+- Updated dependencies [1441041]
+  - @cat-factory/contracts@0.205.0
+  - @cat-factory/kernel@0.204.0
+
+## 0.12.134
+
+### Patch Changes
+
+- Updated dependencies [0b52df7]
+  - @cat-factory/contracts@0.204.0
+  - @cat-factory/kernel@0.203.0
+
 ## 0.12.133
 
 ### Patch Changes